feat(pages): Add discriminate `pageType` to pageContexts
This allows us to create types that will be able to discriminate on this
type to provide correct types for the pageContext.

This change also includes the type stubs, which have yet to be
completely filled out based on what is being spread into the context
objects for these pages.

In a future change we will be using this to determine which Sidebar to
render in the page layout, which will be moved to the wrapPageElement
gatsby API.